+++ crates/miden-tx/src/tests/kernel_tests/test_asset_vault.rs
@@ -7,7 +7,8 @@ use miden_lib::{
         ERR_VAULT_NON_FUNGIBLE_ASSET_ALREADY_EXISTS,
         ERR_VAULT_NON_FUNGIBLE_ASSET_TO_REMOVE_NOT_FOUND,
     },
-    transaction::memory,
+    transaction::{TransactionKernel, memory},
+    utils::word_to_masm_push_string,
 };
 use miden_objects::{
     AssetVaultError,
@@ -23,7 +24,7 @@ use miden_objects::{
 };
 use vm_processor::ProcessState;
 
-use super::{Felt, ONE, Word, ZERO, word_to_masm_push_string};
+use super::{Felt, ONE, Word, ZERO};
 use crate::{
     assert_execution_error, testing::TransactionContextBuilder,
     tests::kernel_tests::read_root_mem_word,
@@ -52,7 +53,12 @@ fn test_get_balance() {
         suffix = faucet_id.suffix(),
     );
 
-    let process = tx_context.execute_code(&code).unwrap();
+    let process = tx_context
+        .execute_code_with_assembler(
+            &code,
+            TransactionKernel::testing_assembler_with_mock_account(),
+        )
+        .unwrap();
 
     assert_eq!(
         process.stack.get(0).as_int(),
@@ -80,7 +86,10 @@ fn test_get_balance_non_fungible_fails() {
         suffix = faucet_id.suffix(),
     );
 
-    let process = tx_context.execute_code(&code);
+    let process = tx_context.execute_code_with_assembler(
+        &code,
+        TransactionKernel::testing_assembler_with_mock_account(),
+    );
 
     assert_execution_error!(
         process,
@@ -111,7 +120,12 @@ fn test_has_non_fungible_asset() {
         non_fungible_asset_key = word_to_masm_push_string(&non_fungible_asset.into())
     );
 
-    let process = tx_context.execute_code(&code).unwrap();
+    let process = tx_context
+        .execute_code_with_assembler(
+            &code,
+            TransactionKernel::testing_assembler_with_mock_account(),
+        )
+        .unwrap();
 
     assert_eq!(process.stack.get(0), ONE);
 }
@@ -147,7 +161,12 @@ fn test_add_fungible_asset_success() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&add_fungible_asset.into())
     );
 
-    let process = &tx_context.execute_code(&code).unwrap();
+    let process = &tx_context
+        .execute_code_with_assembler(
+            &code,
+            TransactionKernel::testing_assembler_with_mock_account(),
+        )
+        .unwrap();
     let process_state: ProcessState = process.into();
 
     assert_eq!(
@@ -190,7 +209,10 @@ fn test_add_non_fungible_asset_fail_overflow() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&add_fungible_asset.into())
     );
 
-    let process = tx_context.execute_code(&code);
+    let process = tx_context.execute_code_with_assembler(
+        &code,
+        TransactionKernel::testing_assembler_with_mock_account(),
+    );
 
     assert_execution_error!(process, ERR_VAULT_FUNGIBLE_MAX_AMOUNT_EXCEEDED);
     assert!(account_vault.add_asset(add_fungible_asset).is_err());
@@ -226,7 +248,12 @@ fn test_add_non_fungible_asset_success() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&add_non_fungible_asset.into())
     );
 
-    let process = &tx_context.execute_code(&code).unwrap();
+    let process = &tx_context
+        .execute_code_with_assembler(
+            &code,
+            TransactionKernel::testing_assembler_with_mock_account(),
+        )
+        .unwrap();
     let process_state: ProcessState = process.into();
 
     assert_eq!(
@@ -264,7 +291,10 @@ fn test_add_non_fungible_asset_fail_duplicate() {
         NON_FUNGIBLE_ASSET = word_to_masm_push_string(&non_fungible_asset.into())
     );
 
-    let process = tx_context.execute_code(&code);
+    let process = tx_context.execute_code_with_assembler(
+        &code,
+        TransactionKernel::testing_assembler_with_mock_account(),
+    );
 
     assert_execution_error!(process, ERR_VAULT_NON_FUNGIBLE_ASSET_ALREADY_EXISTS);
     assert!(account_vault.add_asset(non_fungible_asset).is_err());
@@ -302,7 +332,12 @@ fn test_remove_fungible_asset_success_no_balance_remaining() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&remove_fungible_asset.into())
     );
 
-    let process = &tx_context.execute_code(&code).unwrap();
+    let process = &tx_context
+        .execute_code_with_assembler(
+            &code,
+            TransactionKernel::testing_assembler_with_mock_account(),
+        )
+        .unwrap();
     let process_state: ProcessState = process.into();
 
     assert_eq!(
@@ -343,7 +378,10 @@ fn test_remove_fungible_asset_fail_remove_too_much() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&remove_fungible_asset.into())
     );
 
-    let process = tx_context.execute_code(&code);
+    let process = tx_context.execute_code_with_assembler(
+        &code,
+        TransactionKernel::testing_assembler_with_mock_account(),
+    );
 
     assert_execution_error!(process, ERR_VAULT_FUNGIBLE_ASSET_AMOUNT_LESS_THAN_AMOUNT_TO_WITHDRAW);
 }
@@ -380,7 +418,12 @@ fn test_remove_fungible_asset_success_balance_remaining() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&remove_fungible_asset.into())
     );
 
-    let process = &tx_context.execute_code(&code).unwrap();
+    let process = &tx_context
+        .execute_code_with_assembler(
+            &code,
+            TransactionKernel::testing_assembler_with_mock_account(),
+        )
+        .unwrap();
     let process_state: ProcessState = process.into();
 
     assert_eq!(
@@ -425,7 +468,10 @@ fn test_remove_inexisting_non_fungible_asset_fails() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&non_existent_non_fungible_asset.into())
     );
 
-    let process = tx_context.execute_code(&code);
+    let process = tx_context.execute_code_with_assembler(
+        &code,
+        TransactionKernel::testing_assembler_with_mock_account(),
+    );
 
     assert_execution_error!(process, ERR_VAULT_NON_FUNGIBLE_ASSET_TO_REMOVE_NOT_FOUND);
     assert_matches!(
@@ -462,7 +508,12 @@ fn test_remove_non_fungible_asset_success() {
         FUNGIBLE_ASSET = word_to_masm_push_string(&non_fungible_asset.into())
     );
 
-    let process = &tx_context.execute_code(&code).unwrap();
+    let process = &tx_context
+        .execute_code_with_assembler(
+            &code,
+            TransactionKernel::testing_assembler_with_mock_account(),
+        )
+        .unwrap();
     let process_state: ProcessState = process.into();
 
     assert_eq!(