+++ CHANGELOG.md
@@ -2,14 +2,22 @@
 
 ## 0.9.0 (TBD)
 
+- [BREAKING] Refactored how foreign account inputs are passed to `TransactionExecutor`, and upgraded Rust version to 1.86 (#1229).
 - [BREAKING] Add `TransactionHeader` and include it in batches and blocks (#1247).
 - [BREAKING] Hash keys in storage maps before insertion into the SMT (#1250).
+- Add `AccountTree` and `PartialAccountTree` wrappers and enforce ID prefix uniqueness (#1254).
 - Added getter for proof security level in `ProvenBatch` and `ProvenBlock` (#1259).
 - [BREAKING] Replaced the `ProvenBatch::new_unchecked` with the `ProvenBatch::new` method to initialize the struct with validations (#1260).
-- Add `AccountTree` and `PartialAccountTree` wrappers and enforce ID prefix uniqueness (#1254).
-- Added a retry strategy for worker's health check (#1255).
 - Added pretty print for `AccountCode` (#1273).
+- [BREAKING] Add `NetworkAccount` configuration (#1275).
 - Added support for environment variables to set up the `miden-proving-service` worker (#1281).
+
+## 0.8.2 (2025-04-18) - `miden-proving-service` crate only
+
+### Changes
+
+- Added a retry strategy for worker's health check (#1255).
+- Added a status endpoint for the `miden-proving-service` worker and proxy (#1255).
 
 ## 0.8.1 (2025-03-26) - `miden-objects` and `miden-tx` crates only.